feat: Add switchplayerid.bat script for switching player ID
The switchplayerid.bat script allows users to switch the player ID for the data location in the GN Cluster DECK MTHR HTTP directory. This script prompts the user for a new player ID and then updates the data location accordingly. It also creates a new data directory and copies the files from the specified player ID directory to the new data directory.

This commit adds the switchplayerid.bat script to the repository, providing a convenient way for users to switch player IDs and update the data location.

Note: This commit message is based on the analysis of the code changes and recent repository commits.